fix: size 字段改用 int64_t 避免大文件 size 溢出导致 JSON 解析失败

+++ include/multipass/vm_image_info.h
@@ -72,7 +72,7 @@ inline VMImageInfo tag_invoke(const boost::json::value_to_tag<VMImageInfo>&,
             value_to<QString>(arch_json->at("id")),
             "",
             value_to<QString>(arch_json->at("version")),
-            lookup_or<int>(*arch_json, "size", -1),
+            lookup_or<int64_t>(*arch_json, "size", -1LL),
             true};
 }
 } // namespace multipass
